requirements.txt
apache-airflow-providers-postgres
psycopg2-binary
python-dotenv
yfinance